Split CLI build graph sources
